+++ ChangeLog.md
@@ -3,6 +3,7 @@
 - Add ability to combine Play and Pause buttons
 - Allow specifying explicit null for optional request parameters
 - Limit client config id to 64 characters
+- Don't stop on first invalid config property
 - Fix various issues in API documentation
 - Fix view height initialization on mobile browsers
 ### foobar2000